requirements.txt
fastapi
pydantic
pycrdt
websocket-client
websockets
